# docs/swagger.yaml
basePath: /
definitions:
  handlers.IDsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  models.CardItemFull:
    properties:
      code:
        type: string
      extension:
        type: string
      formula:
        type: string
      id:
        type: integer
      imagePath:
        type: string
      index:
        type: integer
      shared:
        type: boolean
      text:
        type: string
      type:
        $ref: '#/definitions/schema.CardItemType'
      userId:
        type: integer
      width:
        type: string
    type: object
  models.CardItemPartial:
    properties:
      code:
        type: string
      extension:
        type: string
      formula:
        type: string
      id:
        type: integer
      imagePath:
        type: string
      index:
        type: integer
      shared:
        type: boolean
      text:
        type: string
      width:
        type: string
    required:
    - id
    type: object
  models.CardItemShort:
    properties:
      code:
        type: string
      extension:
        type: string
      formula:
        type: string
      imagePath:
        type: string
      index:
        type: integer
      shared:
        type: boolean
      text:
        type: string
      type:
        $ref: '#/definitions/schema.CardItemType'
      width:
        type: string
    required:
    - type
    type: object
...
  schema.CardItemType:
    enum:
    - TEXT
    - TEXT_WITH_HIGHLIGHTED_SYMBOL
    - CODE
    - FORMULA
    - IMAGE
    - WORD_WITH_STRESS
    - MARKDOWN
    type: string
    x-enum-varnames:
    - CardItemTypeText
    - CardItemTypeTextWithHighlightedSymbol
    - CardItemTypeCode
    - CardItemTypeFormula
    - CardItemTypeImage
    - CardItemTypeWordWithStress
    - CardItemTypeMarkdown
...
host: localhost:3033
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Memory nodes, cards, and card items API server.
    **Authentication:** Click **Authorize**, enter your email as username and your password, then Authorize. Swagger will call `POST /users/token` and attach the access token as a Bearer header.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MemoryGuard API
  version: "1.0"
paths:
...
  /card-item/{id}:
    delete:
      description: Deletes a card item by ID owned by the authenticated user
      parameters:
      - description: Card item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Delete card item
      tags:
      - card-items
    get:
      description: Returns a card item by its ID for the authenticated user
      parameters:
      - description: Card item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardItemFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get card item by ID
      tags:
      - card-items
  /card-items:
    get:
      description: Returns all card items for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardItemFull'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List card items
      tags:
      - card-items
...
  /get-card-items:
    post:
      consumes:
      - application/json
      description: Returns multiple card items by their IDs for the authenticated
        user
      parameters:
      - description: List of card item IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardItemFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get card items by IDs
      tags:
      - card-items
...
  /new-card-item:
    post:
      consumes:
      - application/json
      description: Creates a new card item owned by the authenticated user
      parameters:
      - description: Card item
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CardItemShort'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardItemFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Create card item
      tags:
      - card-items
...
  /update-card-item:
    put:
      consumes:
      - application/json
      description: Partially updates an existing card item owned by the authenticated
        user
      parameters:
      - description: Card item update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CardItemPartial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardItemFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Update card item
      tags:
      - card-items
...
schemes:
- http
- https
security:
- Login:
  - api
securityDefinitions:
  Login:
    flow: password
    scopes:
      api: Access the MemoryGuard API
    tokenUrl: /users/token
    type: oauth2
swagger: "2.0"